+++ cmodel/Cmodel/Cmodel/Components/Memory.cpp
@@ -14,7 +14,7 @@ RAM::RAM() {
     memset(ram_, 0, DEFAULT_RAM_SIZE);
 }
 
-RAM::RAM(const uint64_t base, const uint64_t length) {
+RAM::RAM(const uint32_t base, const uint32_t length) {
     this->base_ = base;
     this->size_ = length;
     this->ram_ = new u_char[length];
@@ -23,15 +23,15 @@ RAM::RAM(const uint64_t base, const uint64_t length) {
 
 RAM::~RAM() { delete[] this->ram_; }
 
-const uint64_t& RAM::size() { return size_; }
+const uint32_t& RAM::size() { return size_; }
 
-const uint64_t& RAM::base() { return base_; }
+const uint32_t& RAM::base() { return base_; }
 
-bool RAM::checkRange(uint64_t address) {
+bool RAM::checkRange(uint32_t address) {
     return address >= this->base_ && address < this->base_+size_;
 }
 
-RunCode RAM::readByte(uint64_t address, u_char* dest) {
+RunCode RAM::readByte(uint32_t address, u_char* dest) {
     if (!checkRange(address)) {
         cerr << "Invalid Address" << endl;
         return MemoryViolationExecption;
@@ -40,7 +40,7 @@ RunCode RAM::readByte(uint64_t address, u_char* dest) {
     return Success;
 }
 
-RunCode RAM::readWord(uint64_t address, uint32_t* dest) {
+RunCode RAM::readWord(uint32_t address, uint32_t* dest) {
     if (!checkRange(address)) {
         cerr << "Invalid Address at "<< address << endl;
         return MemoryViolationExecption;
@@ -49,7 +49,7 @@ RunCode RAM::readWord(uint64_t address, uint32_t* dest) {
     return Success;
 }
 
-RunCode RAM::read(uint64_t address, u_char* data, uint64_t len) {
+RunCode RAM::read(uint32_t address, u_char* data, uint32_t len) {
     if (!checkRange(address)) {
         cerr << "Invalid Address" << endl;
         return MemoryViolationExecption;
@@ -58,7 +58,7 @@ RunCode RAM::read(uint64_t address, u_char* data, uint64_t len) {
     return Success;
 }
 
-RunCode RAM::write(uint64_t address, u_char* data, uint64_t len) {
+RunCode RAM::write(uint32_t address, u_char* data, uint32_t len) {
     if (!checkRange(address) || !checkRange(address+len)) {
         cerr << "Invalid Address" << endl;
         return MemoryViolationExecption;
@@ -67,7 +67,7 @@ RunCode RAM::write(uint64_t address, u_char* data, uint64_t len) {
     return Success;
 }
 
-RunCode RAM::writeByte(uint64_t address, u_char data) {
+RunCode RAM::writeByte(uint32_t address, u_char data) {
     if (!checkRange(address)) {
         cerr << "Invalid Address" << endl;
         return MemoryViolationExecption;
@@ -76,7 +76,7 @@ RunCode RAM::writeByte(uint64_t address, u_char data) {
     return Success;
 }
 
-RunCode RAM::writeWord(uint64_t address, uint32_t data) {
+RunCode RAM::writeWord(uint32_t address, uint32_t data) {
     if (!checkRange(address)) {
         cerr << "Invalid Address" << endl;
         return MemoryViolationExecption;
@@ -87,7 +87,7 @@ RunCode RAM::writeWord(uint64_t address, uint32_t data) {
 
 void RAM::show() {
     cout<<"-------RAM Content Start-------"<<endl;
-    for(uint64_t i=base_;i<size_;i+=4) {
+    for(uint32_t i=base_;i<size_;i+=4) {
         uint32_t word;
         this->readWord(i, &word);
         cout<<FHEXI(word)<<endl;
